data/producers/requirements.txt
kafka-python
websocket-client
requests
feedparser
python-dotenv
python-dateutil
beautifulsoup4
# hdfs # 나중에 필요하면 주석 해제